+++ packages/babel-preset/react.js
@@ -1,10 +1,13 @@
-module.exports = function shopifyReactPreset(api, options = {}) {
+module.exports = function shopifyReactPreset(
+  api,
+  {
+    pragma = 'React.createElement',
+    pragmaFrag = 'React.Fragment',
+    transformReactConstantElements = true,
+  } = {},
+) {
   const env = api.env();
 
-  const pragma = options.pragma || 'React.createElement';
-  const pragmaFrag = options.pragmaFrag || 'React.Fragment';
-  const transformReactConstantElements =
-    options.transformReactConstantElements ?? true;
   const plugins = [];
 
   if (env === 'production' && transformReactConstantElements) {